refactor: move spawn-related logic in dedicated component
We will need this logic to provide the user the capability to spawn a single process without backend.
