+++ components/spawn/node/spawn.mjs
